modularize QualityEvaluation
